refactor: move phile.PySide2 to QtCore submodule
This is to mimic PySide2 module's behaviour
of having everything inside a submodule of the PySide2 module.
