+++ phile/notify/gui.py
@@ -24,7 +24,7 @@ import watchdog.observers
 
 # Internal packages.
 import phile
-import phile.PySide2
+import phile.PySide2.QtCore
 import phile.PySide2.QtNetwork
 import phile.data
 import phile.notify
@@ -296,7 +296,8 @@ class MainWindow(QMainWindow):
                     configuration=configuration
                 ),
                 path_handler=functools.partial(
-                    phile.PySide2.call_soon_threadsafe, sorter.update
+                    phile.PySide2.QtCore.call_soon_threadsafe,
+                    sorter.update,
                 ),
                 watched_path=configuration.notification_directory,
                 watching_observer=watching_observer,
@@ -331,7 +332,7 @@ class MainWindow(QMainWindow):
             phile.watchdog.Scheduler(
                 path_filter=entry_point.check_path,
                 path_handler=functools.partial(
-                    phile.PySide2.call_soon_threadsafe,
+                    phile.PySide2.QtCore.call_soon_threadsafe,
                     entry_point.activate_trigger
                 ),
                 watched_path=entry_point.trigger_directory,
