+++ test_phile/test_PySide2/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """
-.. automodule:: test_phile.test_PySide2.test_init
+.. automodule:: test_phile.test_PySide2.test_QtCore
 .. automodule:: test_phile.test_PySide2.test_QtGui
 .. automodule:: test_phile.test_PySide2.test_QtNetwork
 .. automodule:: test_phile.test_PySide2.test_QtWidgets
